refact: adicionando melhorias

+++ src/core/models/sale.py
@@ -1,11 +1,12 @@
 from dataclasses import dataclass
 from datetime import date
+from decimal import Decimal
 from typing import Optional
 
 
 @dataclass
 class Sale:
     product: str
-    price: float
+    price: Decimal
     quantity: int
     data: Optional[date] = None
